User script to redirect to NPR's text version
Since the site doesn't provide a function to do this, make a mechanism
to do it automatically.

The text mode version is light and doesn't track visitors.

In order to redirect to the text version of the article, we pull out the
article's ID and use it in the text mode URL.

Thanks to these MDN docs explaining how to parse query string arguments:
- https://developer.mozilla.org/en-US/docs/Web/API/URL/searchParams#Example
- https://developer.mozilla.org/en-US/docs/Web/API/URLSearchParams/get#Example